requirements.txt
# HTTP客户端
httpx>=0.27.0

# 串口通信
pyserial>=3.5

# AI服务
replicate>=0.25.0

# 配置管理
python-dotenv>=1.0.0

# 重试机制
tenacity>=8.2.3

# 图像处理
Pillow>=10.0.0

# GPIO控制（仅树莓派）
# 注意: picamera2 需要通过 apt 安装: sudo apt-get install -y python3-picamera2
RPi.GPIO; sys_platform == "linux"
gpiozero; sys_platform == "linux"
